requirements.txt
# ------------------------------
# Core Backend Framework
# ------------------------------
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.0
aiofiles==24.1.0

# ------------------------------
# Google Cloud SDKs
# ------------------------------
google-cloud-documentai==2.27.0
google-api-core==2.20.0
google-cloud-storage==2.18.2

# ------------------------------
# Image Processing & OCR
# ------------------------------
Pillow==10.4.0
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
numpy==1.26.4

# ------------------------------
# PDF Generation
# ------------------------------
reportlab==4.2.2

# ------------------------------
# Utilities
# ------------------------------
tqdm==4.66.5
python-dateutil==2.9.0.post0

# ------------------------------
# Optional / Debugging
# ------------------------------
requests==2.32.3
rich==13.9.1
